arm/test: fix printing formats in tests
Fixes:

cow01.c: In function 'parent_check':
../lib/zdtmtst.h:120:11: error: format '%lx' expects argument of type 'long unsigned int', but argument 7 has type 'uint64_t {aka long long unsigned int}' [-Werror=format=]
  test_msg("FAIL: %s:%d: " format " (errno = %d (%s))\n", \
           ^
cow01.c:287:5: note: in expansion of macro 'fail'
     fail("%s[%#x]: %p is not COW-ed (pagemap of "
     ^~~~
In file included from inotify_system.c:14:0:
inotify_system.c: In function 'read_set':
../lib/zdtmtst.h:117:11: error: format '%lu' expects argument of type 'long unsigned int', but argument 5 has type 'unsigned int' [-Werror=form
at=]
  test_msg("ERR: %s:%d: " format " (errno = %d (%s))\n", \
           ^
inotify_system.c:299:3: note: in expansion of macro 'pr_perror'
   pr_perror("read(%d, buf, %lu) Failed, errno=%d",
   ^~~~~~~~~
deleted_dev.c:53:36: error: format '%lx' expects argument of type 'long unsigned int', but argument 4 has type '__dev_t {aka long long unsigned
 int}' [-Werror=format=]
   test_msg("mode %x want %x, dev %lx want %lx\n",
                                    ^
deleted_dev.c:53:45: error: format '%lx' expects argument of type 'long unsigned int', but argument 5 has type 'dev_t {aka long long unsigned i
nt}' [-Werror=format=]
   test_msg("mode %x want %x, dev %lx want %lx\n",
                                             ^
../lib/zdtmtst.h:117:11: error: format '%lx' expects argument of type 'long unsigned int', but argument 5 has type 'off64_t {aka long long int
' [-Werror=format=]
  test_msg("ERR: %s:%d: " format " (errno = %d (%s))\n", \
           ^

Nothing really interesting, but printings with right format specifier.

+++ test/zdtm/static/autofs.c
@@ -188,9 +188,10 @@ static int check_fd(struct autofs_params *p)
 	}
 
 	if (st.st_dev != p->fd_stat.st_dev) {
-		skip("%s: st_dev differs: %lu != %lu "
+		skip("%s: st_dev differs: %llu != %llu "
 		     "(waiting for \"device namespaces\")", p->mountpoint,
-				st.st_dev, p->fd_stat.st_dev);
+				(long long unsigned)st.st_dev,
+				(long long unsigned)p->fd_stat.st_dev);
 //		ret++;
 	}
 	if (st.st_mode != p->fd_stat.st_mode) {
@@ -200,7 +201,7 @@ static int check_fd(struct autofs_params *p)
 	}
 	if (st.st_nlink != p->fd_stat.st_nlink) {
 		pr_err("%s: st_nlink differs: %ld != %ld\n", p->mountpoint,
-				st.st_nlink, p->fd_stat.st_nlink);
+				(long)st.st_nlink, (long)p->fd_stat.st_nlink);
 		ret++;
 	}
 	if (st.st_uid != p->fd_stat.st_uid) {
@@ -214,8 +215,9 @@ static int check_fd(struct autofs_params *p)
 		ret++;
 	}
 	if (st.st_rdev != p->fd_stat.st_rdev) {
-		pr_err("%s: st_rdev differs: %ld != %ld\n", p->mountpoint,
-				st.st_rdev, p->fd_stat.st_rdev);
+		pr_err("%s: st_rdev differs: %lld != %lld\n", p->mountpoint,
+				(long long)st.st_rdev,
+				(long long)p->fd_stat.st_rdev);
 		ret++;
 	}
 	if (st.st_size != p->fd_stat.st_size) {
@@ -484,7 +486,7 @@ static int automountd_loop(int pipe, const char *mountpoint, struct autofs_param
 			return -EINVAL;
 		}
 		if (bytes != sizeof(*packet)) {
-			pr_err("read less than expected: %ld\n", bytes);
+			pr_err("read less than expected: %zd\n", bytes);
 			return -EINVAL;
 		}
 		err = automountd_serve(mountpoint, param, packet);
