+++ test/zdtm/static/inotify_system.c
@@ -297,7 +297,7 @@ int read_set(int inot_fd, char *event_set) {
 	int len;
 	if ((len = read(inot_fd, event_set, EVENT_BUF_LEN)) < 0) {
 		pr_perror("read(%d, buf, %lu) Failed, errno=%d",
-			inot_fd, EVENT_BUF_LEN, errno);
+			inot_fd, (unsigned long)EVENT_BUF_LEN, errno);
 		return -1;
 	}
 	return len;
