+++ test/zdtm/static/inotify_system_nodel.c
@@ -296,7 +296,8 @@ next_event:
 int read_set(int inot_fd, char *event_set) {
 	int len;
 	if ((len = read(inot_fd, event_set, EVENT_BUF_LEN)) < 0) {
-		pr_perror("read(%d, buf, %lu) Failed", inot_fd, EVENT_BUF_LEN);
+		pr_perror("read(%d, buf, %lu) Failed",
+			inot_fd, (unsigned long)EVENT_BUF_LEN);
 		return -1;
 	}
 	return len;
