+++ test/zdtm/static/mntns_link_remap.c
@@ -118,7 +118,8 @@ int ns_child(void *_arg)
 #else
 	if (st1.st_nlink != 0) {
 #endif
-		pr_perror("Wrong number of links: %lu", st1.st_nlink);
+		pr_perror("Wrong number of links: %lu",
+			(long unsigned)st1.st_nlink);
 		exit(1);
 	}
